+++ mobile/android/android-components/components/browser/icons/src/main/java/mozilla/components/browser/icons/Icon.kt
@@ -25,6 +25,11 @@ data class Icon(
         /**
          * This icon was generated.
          */
-        GENERATOR
+        GENERATOR,
+
+        /**
+         * This icon was downloaded.
+         */
+        DOWNLOAD
     }
 }